split device file
